+++ src/Model/ResetPasswordToken.php
@@ -25,10 +25,25 @@ final class ResetPasswordToken
      */
     private $expiresAt;
 
-    public function __construct(string $token, \DateTimeInterface $expiresAt)
+    /**
+     * @var int|null timestamp when the token was created
+     */
+    private $generatedAt;
+
+    /**
+     * @var int expiresAt translator interval
+     */
+    private $transInterval = 0;
+
+    public function __construct(string $token, \DateTimeInterface $expiresAt, int $generatedAt = null)
     {
         $this->token = $token;
         $this->expiresAt = $expiresAt;
+        $this->generatedAt = $generatedAt;
+
+        if (null === $generatedAt) {
+            $this->triggerDeprecation();
+        }
     }
 
     /**
@@ -46,5 +61,83 @@ final class ResetPasswordToken
     public function getExpiresAt(): \DateTimeInterface
     {
         return $this->expiresAt;
+    }
+
+    /**
+     * Get the translation message for when a token expires.
+     *
+     * This is used in conjunction with getExpirationMessageData() method.
+     * Example usage in a Twig template:
+     *
+     * <p>{{ components.expirationMessageKey|trans(components.expirationMessageData) }}</p>
+     *
+     * symfony/translation is required to translate into a non-English locale.
+     *
+     * @throws \LogicException
+     */
+    public function getExpirationMessageKey(): string
+    {
+        $interval = $this->getExpiresAtIntervalInstance();
+
+        switch ($interval) {
+            case $interval->y > 0:
+                $this->transInterval = $interval->y;
+
+                return '%count% year|%count% years';
+            case $interval->m > 0:
+                $this->transInterval = $interval->m;
+
+                return '%count% month|%count% months';
+            case $interval->d > 0:
+                $this->transInterval = $interval->d;
+
+                return '%count% day|%count% days';
+            case $interval->h > 0:
+                $this->transInterval = $interval->h;
+
+                return '%count% hour|%count% hours';
+            default:
+                $this->transInterval = $interval->i;
+
+                return '%count% minute|%count% minutes';
+        }
+    }
+
+    /**
+     * @throws \LogicException
+     */
+    public function getExpirationMessageData(): array
+    {
+        $this->getExpirationMessageKey();
+
+        return ['%count%' => $this->transInterval];
+    }
+
+    /**
+     * Get the interval that the token is valid for.
+     *
+     * @throws \LogicException
+     *
+     * @psalm-suppress PossiblyFalseArgument
+     */
+    public function getExpiresAtIntervalInstance(): \DateInterval
+    {
+        if (null === $this->generatedAt) {
+            throw new \LogicException(\sprintf('%s initialized without setting the $generatedAt timestamp.', self::class));
+        }
+
+        $createdAtTime = \DateTimeImmutable::createFromFormat('U', (string) $this->generatedAt);
+
+        return $this->expiresAt->diff($createdAtTime);
+    }
+
+    private function triggerDeprecation(): void
+    {
+        trigger_deprecation(
+            'symfonycasts/reset-password-bundle',
+            '1.2',
+            'Initializing the %s without setting the "$generatedAt" constructor argument is deprecated. The default "null" will be removed in the future.',
+            self::class
+        );
     }
 }