+++ CHANGELOG.md
@@ -5,11 +5,14 @@ find a change that break's semver, please create an issue.*
 
 ## NEXT
 
+- [#135](https://github.com/SymfonyCasts/reset-password-bundle/pull/134) Add translation support for signature expiration time
+- [#135](https://github.com/SymfonyCasts/reset-password-bundle/pull/134) Fixed invalid signature expiration time
+
 ## v1.2.0
 
 *Dec 10th, 2020*
 
-- #134 - Allow the bundle to be used with PHP 8 - thanks to @ker0x
+- [#134](https://github.com/SymfonyCasts/reset-password-bundle/pull/134)  - Allow the bundle to be used with PHP 8 - thanks to @ker0x
 
 ## v1.1.0